Refactor renderer surface handling and remove unused render_thread module
